refactor: 将排序配置拆分到config模块中，并精简重复代码
