+++ README.md
@@ -122,7 +122,7 @@ chmod +x ./start.sh
 
 #### 索引/面板管理
 - `/构建索引` - 对当前论坛频道构建索引
-- `/创建公开全局搜索` - 创建全局搜索按钮
+- `/创建公开全局搜索` - 创建全局搜索面板
 - `/创建频道搜索` - 在当前帖子内创建频道搜索按钮
 
 #### 算法配置
